requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
requests>=2.31.0
google-cloud-firestore>=2.13.0
google-auth>=2.23.0
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-dateutil>=2.8.0
pytz>=2023.3
structlog>=23.0.0

# AI Orchestration - LangGraph for state machine agents
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Image processing for OCR (payment screenshots with Gemini Vision)
Pillow>=10.0.0

# Additional utilities
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
PyJWT>=2.8.0
email-validator>=2.0.0
python-slugify>=8.0.0